Broke off Ground from Movement and moved it to its own file
